adding application icon cleanup imports
