+++ examples/dynamic_plug_names.py
@@ -1,8 +1,12 @@
 import sys
-from Qt import QtWidgets
+from pathlib import Path
 
 from flowpipe import Graph, INode, InputPlug, OutputPlug
+from Qt import QtWidgets, QtGui
+
 from flowpipe_editor.flowpipe_editor_widget import FlowpipeEditorWidget
+
+BASE_PATH = Path(__file__).parent.parent.resolve()
 
 def compute_embeddings(image):
     """A mock function for a call to a deep learning model or a web service."""
@@ -91,6 +95,7 @@ if __name__ == "__main__":
     
     # Display the graph
     app = QtWidgets.QApplication(sys.argv)
+    app.setWindowIcon(QtGui.QIcon(str(Path(BASE_PATH, 'flowpipe_editor', 'icons', 'flowpipe.png'))))
 
     window = QtWidgets.QWidget()
     window.setWindowTitle("Flowpipe-Editor Dynamic Plug Names Example")
