+++ examples/nested_graphs.py
@@ -1,10 +1,13 @@
 """Nested graphs are supported in flowpipe."""
 import sys
-from Qt import QtWidgets
+from pathlib import Path
 
 from flowpipe import Graph, Node
+from Qt import QtGui, QtWidgets
+
 from flowpipe_editor.flowpipe_editor_widget import FlowpipeEditorWidget
 
+BASE_PATH = Path(__file__).parent.parent.resolve()
 
 @Node(outputs=["file"])
 def MyNode(file):
@@ -62,7 +65,8 @@ if __name__ == "__main__":
 
     # Display the graph
     app = QtWidgets.QApplication(sys.argv)
-
+    app.setWindowIcon(QtGui.QIcon(str(Path(BASE_PATH, 'flowpipe_editor', 'icons', 'flowpipe.png'))))
+    
     window = QtWidgets.QWidget()
     window.setWindowTitle("Flowpipe-Editor Nestd Graphs Example")
     window.resize(1100, 800)
