+++ examples/workflow_design_pattern.py
@@ -1,9 +1,13 @@
 import getpass
 import sys
-from Qt import QtWidgets
+from pathlib import Path
 
 from flowpipe import Graph, Node
+from Qt import QtGui, QtWidgets
+
 from flowpipe_editor.flowpipe_editor_widget import FlowpipeEditorWidget
+
+BASE_PATH = Path(__file__).parent.parent.resolve()
 
 class Workflow(object):
     """Abstract base class defining a workflow, based on a flowpipe graph.
@@ -100,7 +104,8 @@ if __name__ == "__main__":
     
     # Display the graph
     app = QtWidgets.QApplication(sys.argv)
-
+    app.setWindowIcon(QtGui.QIcon(str(Path(BASE_PATH, 'flowpipe_editor', 'icons', 'flowpipe.png'))))
+    
     window = QtWidgets.QWidget()
     window.setWindowTitle("Flowpipe-Editor Workflow Design Pattern Example")
     window.resize(1100, 800)
