+++ examples/world_clock.py
@@ -1,10 +1,14 @@
+import sys
 from datetime import datetime
 from time import time
-from Qt import QtWidgets
-import sys
+from pathlib import Path
 
 from flowpipe import Graph, INode, InputPlug, Node, OutputPlug
+from Qt import QtGui,QtWidgets
+
 from flowpipe_editor.flowpipe_editor_widget import FlowpipeEditorWidget
+
+BASE_PATH = Path(__file__).parent.parent.resolve()
 
 @Node(outputs=["time"])
 def CurrentTime():
@@ -62,6 +66,7 @@ if __name__ == "__main__":
 
     # Display the graph
     app = QtWidgets.QApplication(sys.argv)
+    app.setWindowIcon(QtGui.QIcon(str(Path(BASE_PATH, 'flowpipe_editor', 'icons', 'flowpipe.png'))))
 
     window = QtWidgets.QWidget()
     window.setWindowTitle("Flowpipe-Editor World Clock Example")
